zoom hotfix now only applied for devices with i915 driver loaded

+++ install-zoom.sh
@@ -1,11 +1,17 @@
 #! /bin/bash
 set -e
 
-sudo rm /usr/bin/zoom
+if lsmod | grep "i915" &> /dev/null ; then
+  sudo rm /usr/bin/zoom
+  
+  cat <<"EOF" > /tmp/zoom
+  #!/bin/sh
+  vblank_mode=0 /opt/zoom/ZoomLauncher "$@"
+  EOF
+  
+  sudo install -m 0755 /tmp/zoom /usr/bin
 
-cat <<"EOF" > /tmp/zoom
-#!/bin/sh
-vblank_mode=0 /opt/zoom/ZoomLauncher "$@"
-EOF
+  echo "Added hotfix for i915 driver that disabled vsync"
+fi
 
-sudo install -m 0755 /tmp/zoom /usr/bin
+
